move StringSet code into pkg/skaffold/util
