+++ pkg/skaffold/deploy/deploy_mux.go
@@ -23,8 +23,8 @@ import (
 	"strings"
 
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/build"
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/deploy/util"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubernetes/manifest"
+	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/util"
 )
 
 // DeployerMux forwards all method calls to the deployers it contains.
